refactor code to random-and-replace approach
