--- cfg/multilayer.json
@@ -1,233 +1,0 @@
-{
-    "name": "KnowledgeDistillation",
-    "n_gpu": 1,
-    "teacher": {
-        "type": "DeepWV3Plus",
-        "args": {
-            "num_classes": 19
-        },
-        "snapshot": "checkpoints/cityscapes_best.pth"
-    },
-    "transforms": {
-        "joint_transforms": {
-            "crop_size": 512,
-            "scale_min": 0.5,
-            "scale_max": 2,
-            "ignore_label": 255
-        },
-        "extended_transforms": {
-            "color_aug": 0.2,
-            "blur": "gaussian"
-        }
-    },
-    "train_data_loader": {
-        "type": "CityscapesDataloader",
-        "args": {
-            "data_dir": "data/",
-            "batch_size": 4,
-            "shuffle": true,
-            "validation_split": 0.0,
-            "num_workers": 0,
-            "split": "train",
-            "mode": "fine",
-            "target_type": "semantic",
-            "num_samples": 2048
-        }
-    },
-    "val_data_loader": {
-        "type": "CityscapesDataloader",
-        "args": {
-            "data_dir": "data/",
-            "batch_size": 1,
-            "shuffle": false,
-            "validation_split": 0,
-            "num_workers": 0,
-            "split": "val",
-            "mode": "fine",
-            "target_type": "semantic",
-            "num_samples": 50
-        }
-    },
-    "test_data_loader": {
-        "type": "CityscapesDataloader",
-        "args": {
-            "data_dir": "data/",
-            "batch_size": 1,
-            "shuffle": false,
-            "validation_split": 0,
-            "num_workers": 0,
-            "split": "test",
-            "mode": "fine",
-            "target_type": "semantic",
-            "num_samples": 50
-        }
-    },
-    "optimizer": {
-        "type": "RAdam",
-        "args": {
-            "lr": 0.01
-        }
-    },
-    "supervised_loss": {
-        "type": "CrossEntropyLoss2d",
-        "args": {
-            "ignore_index": 255
-        }
-    },
-    "kd_loss": {
-        "type": "MSELoss",
-        "args": {
-            "reduction": "mean",
-            "num_classes": 1
-        }
-    },
-    "hint_loss": {
-        "type": "MSELoss",
-        "args": {
-            "reduction": "mean",
-            "num_classes": 1000
-        }
-    },
-    "metrics": [],
-    "lr_scheduler": {
-        "type": "MyReduceLROnPlateau",
-        "args": {
-            "mode": "min",
-            "threshold": 0.02,
-            "factor": 0.1,
-            "patience": 0,
-            "verbose": true,
-            "min_lr": 1e-05,
-            "threshold_mode": "rel"
-        }
-    },
-    "trainer": {
-        "name": "ATAKDPTrainer",
-        "epochs": 1000,
-        "save_dir": "saved/",
-        "save_period": 1,
-        "verbosity": 2,
-        "monitor": "min val_mIoU",
-        "early_stop": 1000,
-        "accumulation_steps": 1,
-        "log_step": 5,
-        "do_validation_interval": 4,
-        "lr_scheduler_step_interval": 15,
-        "len_epoch": 200,
-        "tensorboard": true
-    },
-    "pruning": {
-        "pruner": {
-            "dilation": 12,
-            "padding": 12,
-            "kernel_size": 3
-        },
-        "compress_rate": 0.2,
-        "auxiliary_num_layers": 6,
-        "unfreeze": [
-          "mod4.block2",
-          "mod4.block3",
-          "mod4.block4",
-          "mod4.block5",
-          "mod4.block6",
-          "mod5.block2",
-          "mod7.block1.convs.conv2",
-          "mod7.block1.convs.conv3",
-          "aspp.features"
-        ],
-        "hint": [
-          "mod5.block1.convs.conv1",
-          "mod7.block1.convs.conv3",
-          "aspp.features"
-        ],
-        "pruning_plan": [
-            {
-                "name": "mod4.block2.convs.conv2",
-                "compress_rate": 0,
-                "epoch": 1
-            },
-            {
-                "name": "mod4.block3.convs.conv1",
-                "compress_rate": 0,
-                "epoch": 1
-            },
-            {
-                "name": "mod4.block3.convs.conv2",
-                "compress_rate": 0,
-                "epoch": 1
-            },
-            {
-                "name": "mod4.block4.convs.conv2",
-                "compress_rate": 0,
-                "epoch": 1
-            },
-            {
-                "name": "mod4.block5.convs.conv2",
-                "compress_rate": 0,
-                "epoch": 1
-            },
-            {
-                "name": "mod4.block6.convs.conv2",
-                "compress_rate": 0,
-                "epoch": 1
-            },
-            {
-                "name": "mod5.block2.convs.conv2",
-                "compress_rate": 0,
-                "epoch": 1
-            },
-            {
-                "name": "mod7.block1.convs.conv2",
-                "compress_rate": 0,
-                "epoch": 1
-            },
-            {
-                "name": "aspp.features.1.0",
-                "compress_rate": 0,
-                "epoch": 1
-            },
-            {
-                "name": "aspp.features.2.0",
-                "compress_rate": 0,
-                "epoch": 1
-            },
-            {
-                "name": "aspp.features.3.0",
-                "compress_rate": 0,
-                "epoch": 1
-            }
-        ]
-    },
-    "teaching_assistant": {
-        "interval": 10,
-        "tol": 0.01
-    },
-    "weight_scheduler": {
-        "alpha": {
-            "value": 0.0001,
-            "anneal_rate": 2,
-            "max": 0
-        },
-        "beta": {
-            "value": 0.99,
-            "anneal_rate": 0.95,
-            "min": 0.99
-        },
-        "gamma": {
-            "value": 1,
-            "anneal_rate": 1
-        }
-    },
-    "test": {
-        "type": "sliding",
-        "args": {
-            "scales": [1.0],
-            "crop_size": 1024
-        }
-    },
-    "submission": {
-        "save_output": true,
-        "path_output": "submission/",
-        "ext": "png"
-    }
-}
