+++ models/students/__init__.py
@@ -1,4 +1,1 @@
-from .aux_student import AuxStudent
-from .base_student import BaseStudent
-from .independent_student import IndependentStudent
-from .wrapped_student import WrappedStudent
+from .depthwise_student import WrappedStudent
