+++ models/students/transform_blocks/identity.py
@@ -1,0 +1,9 @@
+from torch import nn 
+
+
+class IdentityBlock(nn.Module):
+    def __init__(self):
+        super().__init__()
+
+    def forward(self, x):
+        return x
