--- pruning/__init__.py
@@ -1,1 +1,0 @@
-from .PFEC import PFEC
