+++ train.py
@@ -7,11 +7,10 @@ import losses as module_loss
 import models.metric as module_metric
 import models as module_arch
 import utils.optim as module_optim
-from models.students import BaseStudent, AuxStudent, IndependentStudent, WrappedStudent
+from models.students import WrappedStudent
 from data_loader import _create_transform
 from parse_config import ConfigParser
-from trainer import KDPTrainer, TAKDPTrainer, ATAKDPTrainer, LayerCompressibleTrainer, IndependentTrainer, LayerwiseTrainer
-from pruning import PFEC
+from trainer import LayerwiseTrainer
 from utils import WeightScheduler
 
 # fix random seeds for reproducibility
@@ -36,18 +35,10 @@ def main(config):
     teacher = config.restore_snapshot('teacher', module_arch)
     teacher = teacher.cpu()  # saved some memory as student network will use a (deep) copy of teacher model
 
-    # build models architecture, then print to console
-    args = []
-    if config['trainer']['name'] == 'IndependentTrainer':
-        aux_args = []
-        student = IndependentStudent(teacher, args, aux_args)
-    elif config['trainer']['name'] == 'LayerwiseTrainer':
+    if config['trainer']['name'] == 'LayerwiseTrainer':
         student = WrappedStudent(teacher, config)
-    elif config['trainer']['name'] != 'ATAKDPTrainer':
-        student = BaseStudent(teacher, args)
     else:
-        aux_args = []
-        student = AuxStudent(teacher, args, aux_args)
+        raise NotImplementedError("Supported: Layerwise Trainer")
 
     # get function handles of loss and metrics
     supervised_criterion = config.init_obj('supervised_loss', module_loss)
@@ -67,24 +58,7 @@ def main(config):
         trainer = LayerwiseTrainer(student, criterions, metrics, optimizer, config, train_data_loader,
                                    valid_data_loader, lr_scheduler, weight_scheduler)
     else:
-        pruner = PFEC(student, config)
-        if config['trainer']['name'] == 'LayerCompressibleTrainer':
-            trainer = LayerCompressibleTrainer(student, pruner, criterions, metrics, optimizer, config, train_data_loader,
-                                               valid_data_loader, lr_scheduler, weight_scheduler)
-        elif config['trainer']['name'] == "TAKDPTrainer":
-            trainer = TAKDPTrainer(student, pruner, criterions, metrics, optimizer, config, train_data_loader,
-                                   valid_data_loader, lr_scheduler, weight_scheduler)
-        elif config['trainer']['name'] == 'KDPTrainer':
-            trainer = KDPTrainer(student, pruner, criterions, metrics, optimizer, config, train_data_loader,
-                                 valid_data_loader, lr_scheduler, weight_scheduler)
-        elif config['trainer']['name'] == 'ATAKDPTrainer':
-            trainer = ATAKDPTrainer(student, pruner, criterions, metrics, optimizer, config, train_data_loader,
-                                    valid_data_loader, lr_scheduler, weight_scheduler)
-        elif config['trainer']['name'] == 'IndependentTrainer':
-            trainer = IndependentTrainer(student, pruner, criterions, metrics, optimizer, config, train_data_loader,
-                                         valid_data_loader, lr_scheduler, weight_scheduler)
-        else:
-            raise Exception("Unsupported trainer")
+        raise NotImplementedError("Unsupported trainer")
 
     trainer.train()
