+++ trainer/__init__.py
@@ -1,7 +1,3 @@
 from .kd_trainer import KnowledgeDistillationTrainer
-from .kdp_trainer import KDPTrainer
-from .takdp_trainer import TAKDPTrainer
-from .atakdp import ATAKDPTrainer
 from .layer_compressible import LayerCompressibleTrainer
-from .independent_trainer import IndependentTrainer
 from . layerwise_trainer import LayerwiseTrainer
